move all power tests into one module
